backend/middleware/rate_limiter.py
"""Rate limiting configuration using slowapi."""

from slowapi import Limiter
from slowapi.errors import RateLimitExceeded
from slowapi.util import get_remote_address
from starlette.requests import Request
from starlette.responses import JSONResponse

from backend.config import RATE_LIMIT

limiter = Limiter(key_func=get_remote_address, default_limits=[])

GENERATE_LESSON_LIMIT = RATE_LIMIT


async def rate_limit_exceeded_handler(request: Request, exc: RateLimitExceeded) -> JSONResponse:
    return JSONResponse(
        status_code=429,
        content={"detail": "Too many requests. Please wait before trying again."},
    )
